+++ data-raw/baseline_check.R
@@ -40,12 +40,12 @@ nei <- subset(
 cases <- list(
   counts_hex_duckdb = function() cnefe_counts(CODE_MUNI, h3_resolution = H3_RES, backend = "duckdb", verbose = FALSE),
   counts_hex_r = function() cnefe_counts(CODE_MUNI, h3_resolution = H3_RES, backend = "r", verbose = FALSE),
-  counts_user_duckdb = function() cnefe_counts(CODE_MUNI, polygon_type = "user", polygon = nei, backend = "duckdb", verbose = FALSE),
-  counts_user_r = function() cnefe_counts(CODE_MUNI, polygon_type = "user", polygon = nei, backend = "r", verbose = FALSE),
+  counts_user_duckdb = function() cnefe_counts(CODE_MUNI, polygon = nei, backend = "duckdb", verbose = FALSE),
+  counts_user_r = function() cnefe_counts(CODE_MUNI, polygon = nei, backend = "r", verbose = FALSE),
   lumi_hex_duckdb = function() compute_lumi(CODE_MUNI, h3_resolution = H3_RES, backend = "duckdb", verbose = FALSE),
   lumi_hex_r = function() compute_lumi(CODE_MUNI, h3_resolution = H3_RES, backend = "r", verbose = FALSE),
-  lumi_user_duckdb = function() compute_lumi(CODE_MUNI, polygon_type = "user", polygon = nei, backend = "duckdb", verbose = FALSE),
-  lumi_user_r = function() compute_lumi(CODE_MUNI, polygon_type = "user", polygon = nei, backend = "r", verbose = FALSE),
+  lumi_user_duckdb = function() compute_lumi(CODE_MUNI, polygon = nei, backend = "duckdb", verbose = FALSE),
+  lumi_user_r = function() compute_lumi(CODE_MUNI, polygon = nei, backend = "r", verbose = FALSE),
   tracts_h3 = function() tracts_to_h3(CODE_MUNI, h3_resolution = H3_RES, vars = VARS, verbose = FALSE),
   tracts_polygon = function() tracts_to_polygon(CODE_MUNI, polygon = nei, vars = VARS, verbose = FALSE)
 )